add multitenancy
- update seed (add two companies with one user each)
- reset migrations
- add organizationId filter to all queries/mutations

+++ src/tenants/mutations/deleteTenant.ts
@@ -5,9 +5,8 @@ import { DeleteTenantSchema } from "../schemas"
 export default resolver.pipe(
   resolver.zod(DeleteTenantSchema),
   resolver.authorize(),
-  async ({ id }) => {
-    // TODO: in multi-tenant app, you must add validation to ensure correct tenant
-    const tenant = await db.tenant.deleteMany({ where: { id } })
+  async ({ id }, ctx) => {
+    const tenant = await db.tenant.deleteMany({ where: { id, organizationId: ctx.session.orgId } })
 
     return tenant
   }
